Moves more packets and adds more tests
