+++ Projects/Server/Network/Packets.cs
@@ -35,109 +35,6 @@ namespace Server.Network
     Inspecific = 5
   }
 
-  public sealed class ScrollMessage : Packet
-  {
-    public ScrollMessage(int type, int tip, string text) : base(0xA6)
-    {
-      text ??= "";
-
-      EnsureCapacity(10 + text.Length);
-
-      Stream.Write((byte)type);
-      Stream.Write(tip);
-      Stream.Write((ushort)text.Length);
-      Stream.WriteAsciiFixed(text, text.Length);
-    }
-  }
-
-  public sealed class CurrentTime : Packet
-  {
-    public CurrentTime() : base(0x5B, 4)
-    {
-      var now = DateTime.UtcNow;
-
-      Stream.Write((byte)now.Hour);
-      Stream.Write((byte)now.Minute);
-      Stream.Write((byte)now.Second);
-    }
-  }
-
-  public sealed class MapChange : Packet
-  {
-    public MapChange(Mobile m) : base(0xBF)
-    {
-      EnsureCapacity(6);
-
-      Stream.Write((short)0x08);
-      Stream.Write((byte)(m.Map?.MapID ?? 0));
-    }
-  }
-
-  public sealed class SupportedFeatures : Packet
-  {
-    public SupportedFeatures(NetState ns) : base(0xB9, ns.ExtendedSupportedFeatures ? 5 : 3)
-    {
-      var flags = ExpansionInfo.CoreExpansion.SupportedFeatures;
-
-      flags |= Value;
-
-      if (ns.Account.Limit >= 6)
-      {
-        flags |= FeatureFlags.LiveAccount;
-        flags &= ~FeatureFlags.UOTD;
-
-        if (ns.Account.Limit > 6)
-          flags |= FeatureFlags.SeventhCharacterSlot;
-        else
-          flags |= FeatureFlags.SixthCharacterSlot;
-      }
-
-      if (ns.ExtendedSupportedFeatures)
-        Stream.Write((uint)flags);
-      else
-        Stream.Write((ushort)flags);
-    }
-
-    public static FeatureFlags Value { get; set; }
-
-    public static SupportedFeatures Instantiate(NetState ns) => new SupportedFeatures(ns);
-  }
-
-  public static class AttributeNormalizer
-  {
-    public static int Maximum { get; set; } = 25;
-
-    public static bool Enabled { get; set; } = true;
-
-    public static void Write(PacketWriter stream, int cur, int max)
-    {
-      if (Enabled && max != 0)
-      {
-        stream.Write((short)Maximum);
-        stream.Write((short)(cur * Maximum / max));
-      }
-      else
-      {
-        stream.Write((short)max);
-        stream.Write((short)cur);
-      }
-    }
-
-    public static void WriteReverse(PacketWriter stream, int cur, int max)
-    {
-      if (Enabled && max != 0)
-      {
-        stream.Write((short)(cur * Maximum / max));
-        stream.Write((short)Maximum);
-      }
-      else
-      {
-        stream.Write((short)cur);
-        stream.Write((short)max);
-      }
-    }
-  }
-
   public sealed class MobileHits : Packet
   {
     public MobileHits(Mobile m) : base(0xA1, 9)
